Move WidgetState
