+++ backend-embedded-graphics/src/widgets/label/ascii.rs
@@ -2,7 +2,7 @@ use embedded_graphics::{
     mono_font::{ascii, MonoTextStyle, MonoTextStyleBuilder},
     pixelcolor::PixelColor,
 };
-use embedded_gui::{geometry::BoundingBox, widgets::label::Label, WidgetState};
+use embedded_gui::{geometry::BoundingBox, state::WidgetState, widgets::label::Label};
 
 use crate::{themes::Theme, widgets::label::LabelStyle};
