+++ backend-embedded-graphics/src/widgets/label/latin1.rs
@@ -3,7 +3,7 @@ use embedded_graphics::{
     mono_font::{latin1, MonoTextStyle, MonoTextStyleBuilder},
     pixelcolor::PixelColor,
 };
-use embedded_gui::{geometry::BoundingBox, widgets::label::Label, WidgetState};
+use embedded_gui::{geometry::BoundingBox, state::WidgetState, widgets::label::Label};
 
 use crate::{themes::Theme, widgets::label::LabelStyle};
